+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGBFEPGWE9VP1XZZHJK5JQP"
+        "01KKGJ3CD5ZC3PDDM5VMJVH9P0"
     )
     r"""The unique identifier of the data source"""