+++ packages/orq-rc/src/orq_ai_sdk/models/fileuploadop.py
@@ -134,7 +134,7 @@ class FileUploadResponseBody(BaseModel):
     workspace_id: str
     r"""The id of the resource"""
 
-    created: Optional[datetime] = parse_datetime("2026-03-12T06:24:10.751Z")
+    created: Optional[datetime] = parse_datetime("2026-03-12T08:19:55.554Z")
     r"""The date and time the resource was created"""
 
     @model_serializer(mode="wrap")