+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KKGBFEPKZG5KX7GG1ZDXFW35"
+        "01KKGJ3CD757N4BGHPQXEQC71Y"
     )
     r"""The unique identifier of the data source"""